ido commit cluster.py

+++ app/api/api_v1/routers/cluster.py
@@ -3,6 +3,7 @@ import json
 import pandas as pd
 from fastapi import APIRouter, Query, Request, Depends, Response, encoders
 from typing import List, Optional
+from pathlib import Path
 from app.db.session import get_db
 from app.db.crud import (
     get_genes, get_strains_cluster, get_strain_id_name
@@ -24,12 +25,16 @@ async def cluster_tree(
 ):
     """Get all strains"""
     strains = get_strains_cluster(db,gene_name)
-    cluster_id = strains[0] #cluster id is used for the hash id used for later
+    cluster_id = str(strains[0]) #cluster id is used for the hash id used for later
+    filenameHash = hashlib.md5(cluster_id.encode())
+    filename = filenameHash.hexdigest()
+    my_file = Path(r'static/cluster/'+filename+'.txt')
+    if my_file.is_file():
+        return 1
 
     dict_id_strain = json.loads(strains[1].replace("'", "\""))
     data_id_strain = pd.DataFrame(dict_id_strain.items(), columns=['index', 'count'])
     data_id_strain['index'] = data_id_strain['index'].astype(int)
     complet_data = get_strain_id_name(db,data_id_strain)
-
     complet_data.to_csv(r'static/cluster/cluster.csv', index=False)
     return 0
